cv/api - clean skills

+++ computervision/JobExecuter.py
@@ -2,6 +2,7 @@
 
 import time
 import json
+from collections import defaultdict
 from managers.DataRepository import DataRepository
 from Predictor import Predictor
 from Trainer import Trainer
@@ -18,6 +19,59 @@ REPO = DataRepository()
 no_shutdown_job = True
 predictor = Predictor()
 trainer = Trainer()
+
+def defaultdict_to_dict(d):
+    if isinstance(d, defaultdict):
+        d = dict(d)
+    return {
+        k: defaultdict_to_dict(v) if isinstance(v, (dict, defaultdict)) else v
+        for k, v in d.items()
+    }
+
+def clean_skills(traintest: str):
+    skills = REPO.get_skills(traintest)
+
+    for index, skillrow in skills.iterrows():
+        rowId = skillrow['id']
+        if rowId > 1203:
+            continue
+        print("@"*40, rowId, "@"*40)
+        cleaned_skill = {}
+        uncleaned_skill : dict = skillrow['skillinfo']
+
+        for composition_name, composition_properties in uncleaned_skill.items():
+            cleaned_skill[composition_name] = []
+            for composition_index, composition_stages in enumerate(composition_properties):
+                cleaned_composition_stages = defaultdict(lambda: defaultdict(lambda: defaultdict(dict)))
+                for stage, stageProperties in composition_stages.items():
+                    if stage == 'compositionName':
+                        continue
+                    elif stage != 'StageProperties':
+                        for property, value in stageProperties.items():
+                            if property in ['Feet', 'RopesHolding'] and value != 2:
+                                print(composition_name, composition_index, stage, property, value)
+                                cleaned_composition_stages[stage][property] = value
+                            elif property not in ['Feet', 'TurningBackwards', 'PointWorthy', 'RopesHolding'] and value != 0:
+                                print(composition_name, composition_index, stage, property, value)
+                                cleaned_composition_stages[stage][property] = value
+                            elif (property == 'PointWorthy' and value != 'true') or (property == 'TurningBackwards' and value != 60):
+                                cleaned_composition_stages[stage][property] = value
+                            
+                    else:
+                        for numericStage, numericStageProperties in stageProperties.items():
+                            for property, value in numericStageProperties.items():
+                                if property in ['Feet', 'RopesHolding'] and value != 2:
+                                    print(composition_name, composition_index, stage, numericStage, property, value)
+                                    cleaned_composition_stages[stage][numericStage][property] = value
+                                elif property not in ['Feet', 'TurningBackwards', 'PointWorthy', 'RopesHolding'] and value != 0:
+                                    print(composition_name, composition_index, stage, numericStage, property, value)
+                                    cleaned_composition_stages[stage][numericStage][property] = value
+                                elif (property == 'PointWorthy' and value != 'true') or (property == 'TurningBackwards' and value != 60):
+                                    cleaned_composition_stages[stage][numericStage][property] = value
+
+                cleaned_skill[composition_name].append(defaultdict_to_dict(cleaned_composition_stages))
+
+        REPO.update_skill(skillId=rowId, skillinfo=json.dumps(cleaned_skill))
 
 while no_shutdown_job:
     job = REPO.get_next_job()
@@ -53,6 +107,10 @@ while no_shutdown_job:
             from_scratch=True,
             save_anyway=True,
         )
+    elif job["type"] == "CLEAN":
+        clean_skills('train')
+        clean_skills('val')
+        pass
 
         REPO.delete_job(job["id"])
     else:
